fix: Networking update with N^2 complexity (#2674)
`Networking update with N^2 complexity`: feat(networking)

Currently when a player joins any dimension, the server will launch N
calls of `setShouldSyncBetweenClients`, once per each person connected
to the server.
Once this happens each client will fire a call of `syncClients` which
will fire a packet for each other player.
This sequence will saturate the network with many small calls, causing
short but heavy lag spikes.

This fix was tested on a server with an average population of 20 and a
max of 50 players.
Please tell me if there's any issues.

---

I agree to the Contributor License Agreement (CLA).

---------

Co-authored-by: Aether-Team-Bot <88561924+Aether-Team-Bot@users.noreply.github.com>
Co-authored-by: Juniper <67175453+Ultrasquid9@users.noreply.github.com>
Co-authored-by: Aether-Team-Bot <machine@gildedgames.com>
Co-authored-by: bconlon <bconlon.gml@gmail.com>
Co-authored-by: bconlon <67203206+bconlon1@users.noreply.github.com>