fix: capability not allowing filling by milk (#2753)
Milk wasn't allowed as fluid for filling SkyrootBucketWrapper, and then
afterwards was not actually in the REPLACEMENTS map to actually be
swapped.

Fixes #2752

Relates to #2569 so would be nice for this to be included in #2748 as
well.

---

I agree to the Contributor License Agreement (CLA).

(cherry picked from commit 842d8212b2db1f1169d52d9cc0766cff512fce01)